cmd/compile: enable two orphaned tests
These tests weren't being run.  Re-enable them.

R=go1.12

Reviewed-on: https://go-review.googlesource.com/127117
Run-TryBot: Keith Randall <khr@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
